\name{FBgn2anno}
\alias{FBgn2anno}
\docType{data}
\title{GO annotation of flybase genes}
\description{GO annotation of the 1367 genes targeted in this screen.}
\usage{data(FBgn2anno)}
\value{GO annotation.}
\format{
  A data frame with 118863 observations on the following 4 variables.\cr
  \describe{
    \item{\code{source}}{a character vector}
    \item{\code{gene_id}}{a character vector}
    \item{\code{Category}}{a character vector}
    \item{\code{Name}}{a character vector}
  }
}
\seealso{\link{DmelSGI-package}}
\examples{
data(FBgn2anno)
}
\keyword{datasets}
