\name{orderDim}
\alias{orderDim}
\title{Orders a three dimensional array along one dimension}
\description{Orders a three dimensional array along one dimension.}
\usage{
orderDim(x, i)
}
\arguments{
  \item{x}{A three dimensional array.}
  \item{i}{The array is ordered along the i-th dimension.}
}
\details{The three dimensional array is ordered along the i-th dimension.}
\value{An integer vector with the ordering.}
\author{Bernd Fischer}
\seealso{\link{DmelSGI-package}}
\examples{
print(orderDim)
}
\keyword{ datasets }
