\name{SelectedClusters}
\alias{SelectedClusters}
\docType{data}
\title{Selected processes displayed on the hairball}
\description{A list of gene sets that are displayed on the hairball.}
\usage{data(SelectedClusters)}
\value{A list with selected clusters for visualization.}
\format{
  The format is:\cr
List of 24\cr
 $ SWI/SNF                                 : chr [1:6] "Bap60" "brm" "dalao" ...\cr
 $ Condensin/Cohesin                       : chr [1:5] "SMC2" "Cap-D2" "glu" ...\cr
 $ Cytokinesis                             : chr [1:12] "pav" "sqh" "rok" ...\cr
 ...
}
\seealso{\link{DmelSGI-package}}
\examples{
data(SelectedClusters)
}
\keyword{datasets}
