\name{SelectedClustersComplexes}
\alias{SelectedClustersComplexes}
\docType{data}
\title{Selected complexes}
\description{Manually curated protein complexes.}
\value{A list of gene sets that represent well-defined protein complexes.}
\usage{data(SelectedClustersComplexes)}
\format{
  The format is:\cr
List of 27\cr
 $ DREAM complex                           : chr [1:5] "Caf1" "mip120" "mip130" "mip40" ...\cr
 $ Condensin/Cohesin                       : chr [1:5] "SMC2" "Cap-D2" "glu" "eco" ...\cr
 $ Apc/C                                   : chr [1:9] "APC10" "Cdc23" "Cdc16" "shtd" ...\cr
 ...
}
\seealso{\link{DmelSGI-package}}
\examples{
data(SelectedClustersComplexes)
}
\keyword{datasets}
