\name{toMatrix}
\alias{toMatrix}
\title{Flattens a three dimensional array to a two dimensional matrix}
\description{Flattens a three dimensional array to a two dimensional matrix.}
\usage{
toMatrix(x)
}
\arguments{
  \item{x}{A three dimensional array.}
}
\value{A matrix.}
\seealso{\link{DmelSGI-package}}
\examples{
print(toMatrix)
}
\keyword{ datasets }
