\name{trsf}
\alias{trsf}
\title{Transform a correlation to a distance}
\description{Transforms a correlation to a distance.}
\usage{
trsf(x)
}
\arguments{
  \item{x}{A real valued vector with values in [-1,1]. It is intended to be a correlation.}
}
\value{a real valued vector of distances.}
\seealso{\link{DmelSGI-package}}
\examples{
print(trsf)
}
\keyword{ datasets }
